Made pls say only work for me

+++ Markos.js
@@ -65,9 +65,11 @@ client.on('message', msg => {
       owner.sendMessage(`**${username}#${discriminator} (${id}):**\nThis person was trying to use the eval command!\n${script}`)
     }
   } else if (command === 'say') {
-    const say = args.join(' ')
-    msg.delete();
-    msg.channel.sendMessage(say)
+    if (msg.author.id === config.owner) {
+      const say = args.join(' ')
+      msg.delete();
+      msg.channel.sendMessage(say)
+    }
   } else {
     fs.access(path.join(commandsPath, command + '.js'), fs.constants.R_OK, (err) => {
       if (err) {
@@ -104,7 +106,7 @@ client.on('guildDelete', guild => {
 client.on('ready', () => {
   console.log('Markos ' + config.version + ' loaded successfully. 👌')
   client.user.setGame(config.prefix + 'help 👌 👀')
- // client.guilds.get('281482896265707520').fetchMember("284901153265680385").then((guildMember) => guildMember.setNickname('MemeDaddy Markos')).catch(console.error);
+  // client.guilds.get('281482896265707520').fetchMember("284901153265680385").then((guildMember) => guildMember.setNickname('MemeDaddy Markos')).catch(console.error);
   console.log('Welcome, Austin. 👀')
 })
